Reorganized acceptance test page classes
The page classes are now grouped in sub-packages by system/IDA.
