+++ acceptance_tests/pages/__init__.py
@@ -1,0 +1,4 @@
+# pylint: disable=wildcard-import
+from acceptance_tests.pages.ecommerce import *
+from acceptance_tests.pages.lms import *
+from acceptance_tests.pages.marketing import *
